requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Audio processing
soundfile>=0.12.1
librosa>=0.10.0
pydub>=0.25.1
audiomentations>=0.33.0

# Evaluation metrics
jiwer>=3.0.0
evaluate>=0.4.0

# VAD
silero-vad>=4.0.0

# MLOps & Tracking
wandb>=0.15.0
tensorboard>=2.14.0
mlflow>=2.8.0

# Utilities
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
scipy>=1.11.0

# Fast inference (optional)
# faster-whisper>=0.9.0
# onnxruntime-gpu>=1.16.0

# Jupyter support
ipywidgets>=8.1.0
jupyter>=1.0.0
